Widgets // Move WidgetBackgroundAppEntity impl into BGOptions.
